requirement.txt
Flask==2.2.5
Flask-Mail==0.9.1
bcrypt==4.0.1
langchain==0.0.350
openai==1.14.1
python-dotenv==1.0.1
Werkzeug==2.2.3
itsdangerous==2.1.2
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5



